Change Shell to Term  changing scripts to use make not build.sh where possible stars.bjs and variaous Term fixes start of tui in Inpu classes

+++ binjs/src/libbash/build.sh
@@ -15,11 +15,10 @@
 #	tar xvf bash-4.2.tar
 #	then set BASH_SRC in this script
 
-# Change this to where ever you installed bash
-
 cd `dirname $0`
 SRC=`pwd`
 
+# Change this to where ever you installed bash
 export BASH_SRC=../../../bash-4.2
 
 			
@@ -38,24 +37,21 @@ cp -v ${SRC}/libbash.* .
 
 # you only need to build bash with -fPIC once for dev of libbash.c itself
 if [ "$1" != "--quick" ] ; then
-	make -s clean
-	./configure CFLAGS=-fPIC
-	make -s
+	make -s clean  || exit 2
+	./configure CFLAGS=-fPIC && make -s  || exit 4
 else 
-	make -s
+	make -s  || exit 4
 fi
 
-# make libbash.a
-ar -r libbash.a ${BASSH_OBJECTS}
+test -f libbash.so.1 || ( echo "libbash.so.1 not built" && exit 7 )
 
 rm ${SRC}/../../contrib/libbash.*
 
 cp -v libbash.so.1 libbash.a ${SRC}/../../contrib
 ln -s ${SRC}/../../contrib/libbash.so.1 ${SRC}/../../contrib/libbash.so
-# ls -l $SRC/../../contrib
 
 cd ${SRC}
 gcc -Wall -fPIC  test/test_libbash.c -o test/test_libbash -lbash -L../../contrib
 
-ls -l test/test_libbash
+ls -l test/test_libbash >/dev/null || ( echo "test_libbash not built" && exit 7 )
